requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alembic==1.6.5
    # via apache-airflow
apache-airflow-providers-amazon==1.4.0
    # via
    #   -r requirements.in
    #   apache-airflow
apache-airflow-providers-ftp==1.1.0
    # via apache-airflow
apache-airflow-providers-imap==1.0.1
    # via apache-airflow
apache-airflow-providers-sqlite==1.0.2
    # via apache-airflow
apache-airflow[amazon]==2.1.0
    # via
    #   -r requirements.in
    #   apache-airflow-providers-amazon
apispec[yaml]==3.3.2
    # via flask-appbuilder
argcomplete==1.12.3
    # via apache-airflow
attrs==20.3.0
    # via
    #   apache-airflow
    #   cattrs
    #   jsonschema
babel==2.9.1
    # via flask-babel
blinker==1.4
    # via apache-airflow
boto3==1.17.84
    # via
    #   apache-airflow-providers-amazon
    #   watchtower
botocore==1.20.84
    # via
    #   boto3
    #   s3transfer
cattrs==1.7.1
    # via apache-airflow
certifi==2021.5.30
    # via httpx
cffi==1.14.5
    # via cryptography
click==7.1.2
    # via
    #   clickclick
    #   flask
    #   flask-appbuilder
clickclick==20.10.2
    # via apache-airflow
colorama==0.4.4
    # via
    #   flask-appbuilder
    #   rich
colorlog==5.0.1
    # via apache-airflow
commonmark==0.9.1
    # via rich
croniter==1.0.13
    # via apache-airflow
cryptography==3.4.7
    # via apache-airflow
defusedxml==0.7.1
    # via python3-openid
dill==0.3.3
    # via apache-airflow
dnspython==2.1.0
    # via email-validator
docutils==0.17.1
    # via python-daemon
email-validator==1.1.2
    # via flask-appbuilder
flask-appbuilder==3.3.0
    # via apache-airflow
flask-babel==1.0.0
    # via flask-appbuilder
flask-caching==1.10.1
    # via apache-airflow
flask-jwt-extended==3.25.1
    # via flask-appbuilder
flask-login==0.4.1
    # via
    #   apache-airflow
    #   flask-appbuilder
flask-openid==1.2.5
    # via flask-appbuilder
flask-sqlalchemy==2.5.1
    # via flask-appbuilder
flask-wtf==0.14.3
    # via
    #   apache-airflow
    #   flask-appbuilder
flask==1.1.4
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   flask-babel
    #   flask-caching
    #   flask-jwt-extended
    #   flask-login
    #   flask-openid
    #   flask-sqlalchemy
    #   flask-wtf
graphviz==0.16
    # via apache-airflow
gunicorn==20.1.0
    # via apache-airflow
h11==0.12.0
    # via httpcore
httpcore==0.13.3
    # via httpx
httpx==0.18.1
    # via apache-airflow
idna==3.2
    # via
    #   email-validator
    #   rfc3986
importlib-metadata==1.7.0
    # via apache-airflow
importlib-resources==1.5.0
    # via apache-airflow
inflection==0.5.1
    # via apache-airflow
iso8601==0.1.14
    # via apache-airflow
isodate==0.6.0
    # via openapi-schema-validator
itsdangerous==1.1.0
    # via
    #   apache-airflow
    #   flask
    #   flask-wtf
jinja2==2.11.3
    # via
    #   apache-airflow
    #   flask
    #   flask-babel
    #   python-nvd3
    #   swagger-ui-bundle
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   openapi-schema-validator
    #   openapi-spec-validator
lazy-object-proxy==1.6.0
    # via apache-airflow
lockfile==0.12.2
    # via
    #   apache-airflow
    #   python-daemon
mako==1.1.4
    # via alembic
markdown==3.3.4
    # via apache-airflow
markupsafe==1.1.1
    # via
    #   apache-airflow
    #   jinja2
    #   mako
    #   wtforms
marshmallow-enum==1.5.1
    # via flask-appbuilder
marshmallow-oneofschema==2.1.0
    # via apache-airflow
marshmallow-sqlalchemy==0.23.1
    # via flask-appbuilder
marshmallow==3.12.1
    # via
    #   flask-appbuilder
    #   marshmallow-enum
    #   marshmallow-oneofschema
    #   marshmallow-sqlalchemy
numpy==1.20.3
    # via
    #   apache-airflow
    #   pandas
openapi-schema-validator==0.1.5
    # via openapi-spec-validator
openapi-spec-validator==0.3.1
    # via apache-airflow
pandas==1.2.4
    # via apache-airflow
pendulum==2.1.2
    # via apache-airflow
prison==0.1.3
    # via flask-appbuilder
psutil==5.8.0
    # via apache-airflow
pycparser==2.20
    # via cffi
pygments==2.9.0
    # via
    #   apache-airflow
    #   rich
pyjwt==1.7.1
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   flask-jwt-extended
pyrsistent==0.17.3
    # via jsonschema
python-daemon==2.3.0
    # via apache-airflow
python-dateutil==2.8.1
    # via
    #   alembic
    #   apache-airflow
    #   botocore
    #   croniter
    #   flask-appbuilder
    #   pandas
    #   pendulum
python-editor==1.0.4
    # via alembic
python-nvd3==0.15.0
    # via apache-airflow
python-slugify==4.0.1
    # via
    #   apache-airflow
    #   python-nvd3
python3-openid==3.2.0
    # via
    #   apache-airflow
    #   flask-openid
pytz==2021.1
    # via
    #   babel
    #   flask-babel
    #   pandas
pytzdata==2020.1
    # via pendulum
pyyaml==5.4.1
    # via
    #   apache-airflow
    #   apispec
    #   clickclick
    #   openapi-spec-validator
rfc3986[idna2008]==1.5.0
    # via httpx
rich==10.2.2
    # via apache-airflow
s3transfer==0.4.2
    # via boto3
setproctitle==1.2.2
    # via apache-airflow
six==1.16.0
    # via
    #   flask-jwt-extended
    #   isodate
    #   jsonschema
    #   openapi-schema-validator
    #   openapi-spec-validator
    #   prison
    #   python-dateutil
    #   sqlalchemy-utils
    #   tenacity
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
sqlalchemy-jsonfield==1.0.0
    # via apache-airflow
sqlalchemy-utils==0.37.5
    # via flask-appbuilder
sqlalchemy==1.3.24
    # via
    #   alembic
    #   apache-airflow
    #   flask-appbuilder
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   sqlalchemy-jsonfield
    #   sqlalchemy-utils
swagger-ui-bundle==0.0.8
    # via apache-airflow
tabulate==0.8.9
    # via apache-airflow
tenacity==6.2.0
    # via apache-airflow
termcolor==1.1.0
    # via apache-airflow
text-unidecode==1.3
    # via python-slugify
unicodecsv==0.14.1
    # via apache-airflow
urllib3==1.26.5
    # via botocore
watchtower==0.7.3
    # via apache-airflow-providers-amazon
werkzeug==1.0.1
    # via
    #   apache-airflow
    #   flask
    #   flask-jwt-extended
wtforms==2.3.3
    # via flask-wtf
zipp==3.4.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
